Extract base class from PropertyAssignmentIntegrationTest
This is a preparation to split the test class.
